`Pointer::pop_back`: Correctly handle empty strings
Empty strings weren't correctly handled in `pop_back`: When calling `pop_back`
for the pointer `//`, it would leave an empty string for `Pointer::inner` and
`Pointer::count` at 1. Which is obviously wrong and leaves `Pointer`
inconsistent.

I also synchronized `fn back` with this, although in this case it was correct.