README: Fix `Delete` example
Seems like this was accidentally left as a copy-paste of the `Assign`
example. I took the liberty to add some extra keys to the object which
we're deleting from, as I think it makes it a bit clearer what's
happening than:

    { "secret": { "universe": 42 }} -> { "secret": {} }

Which confused me a bit at first (but that might just be me).